Pull card images from holotable instead of s3
To have a single source of truth and easier image updates,
pull card images from the holotable repo via get-card-images.sh.

Running this script is idempotent and should refresh any updated
files that have changes. Git is required on the VM for this
change and has been added to bootstrap script.