Temporarily disable player IP checks
Disable similar player and league match checks based on IP.

With SSL termination now added via NGINX, the REMOTE_ADDR HTTP
header that is passed to the java process is no longer the client's
IP but instead the server's IP. This is a very common situation,
and most http libraries support the alternate X-Forwarded-For header
for use in a proxy situation. However, the Java netty library that
GEMP uses does not seem to pull this in by default. I tried the
solution discussed here:
https://stackoverflow.com/questions/6520624/jboss-netty-getting-user-ip-http-request
but did not have success initially.

In the meantime, disable ip-based duplicate account logic temporarily.