Ezra Bridger Bug Fix:
- If opponent chooses for you to activate force, changed the minimum you can choose to activate to 0.  Previously it was 1.

I looked into what it would take to make it into a two step process, where:

Step 1: You get asked if you want to activate any force.
Step 2: You pick how much, 1 to X.

But the way I could get it to work was in reverse, where it asks the player to pick 1 to X, then asks if they want to activate
that much force.  That's clearly worse.

I don't think this way has any downsides, is better than the alternative, and is the easiest change.