Store boolean instead of repeatedly querying database
This was querying the database every time it needed to know if private games were enabled. Now it only checks when it starts up and every time an admin toggles the setting. This means that editing the setting in the database won't change anything until restart or an admin toggles it.